use kcookiejar through DBus, not KIO::Integration
With KIO::Integration::CookieJar setting/storing cookies sometimes fail, with
the error message "QNetworkRequest::setHeader: QVariant of type
QList<QNetworkCookie> cannot be used with header Cookie". This commit tries to
fix the problem, but setting cookies still does not work because I cannot find
the right way to pass a number as qlonglong to a DBus method.