list Ruby scripts in stdin instead of in XDG dirs
Instead of looking for scripts to require in specific directories, read a list
of filenames of the scripts in stdin. This makes it more convenient to select
which scripts to load. In the future a launcher can be written to look for
scripts in XDG directories.

Adjust README.rst to reflect the change.

Make icon.rb require 'xdg' since it is no longer required by the main program.
Need to make sure scripts do not rely on requires in other files that might be
removed.